Added a new file and changed some of the visualizations.
